pyproject.toml
[project]
name = "saturn"
authors = [
    { name = "Asofi", email = "abdoalsofi576@gmail.com"}
]
description = "Frappe app saturn"
requires-python = ">=3.10"
readme = "README.md"
dynamic = ["version"]
dependencies = [
    # "frappe~=15.0.0" # Installed and managed by bench.
    "python-docx==1.2.0",
    "qrcode[pil]==7.4.2",
    "Pillow>=9.0.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.2.0"
]

[build-system]
requires = ["flit_core >=3.4,<4"]
build-backend = "flit_core.buildapi"

# These dependencies are only installed when developer mode is enabled
[tool.bench.dev-dependencies]
# package_name = "~=1.1.0"
